refactor structure folder components page
