+++ components/event/EventItem.tsx
