+++ components/layout/MainLayout.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import Header from "components/Header";
-import Footer from "components/Footer";
+import Header from "components/header";
+import Footer from "components/footer";
 
 export default function MainLayout({
   children,
